requirements.txt
streamlit
python-dotenv
google-generativeai
firebase-admin
pyrebase